Better structure for tippecanoe code for use outside of atlas
